fix: fixed bibliography build issue
`apa.bst` is not distributed with TexLive 2020, so the file has to be manually added to this project. The style is from the '90s, and has been superseded by a more modern implementation. When the research was written in 2017, `apa.bst` was still used, so it's copied here to maintain the original manuscript.